# fix(notification): optimize unread notifications query using index