--- a/db/schema.rb
+++ b/db/schema.rb
@@ -10,7 +10,7 @@
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema[7.0].define(version: 2024_10_20_082634) do
+ActiveRecord::Schema[7.0].define(version: 2025_02_17_101253) do
   # These are extensions that must be enabled in order to support this database
   enable_extension "plpgsql"
 
@@ -308,6 +308,7 @@ ActiveRecord::Schema[7.0].define(version: 2024_10_20_082634) do
     t.datetime "created_at", null: false
     t.datetime "updated_at", null: false
     t.index ["read_at"], name: "index_noticed_notifications_on_read_at"
+    t.index ["recipient_type", "recipient_id"], name: "idx_noticed_notifications_unread", where: "(read_at IS NULL)"
     t.index ["recipient_type", "recipient_id"], name: "index_noticed_notifications_on_recipient"
   end