Fixes #5391 

#### Describe the changes you have made in this PR -
Optimized the unread notifications query by adding a more efficient index. This optimization improves performance by reducing the execution time when fetching unread notifications.

### Screenshots of the changes (If any) -
![Screenshot from 2025-02-17 15-35-13](https://github.com/user-attachments/assets/a3f7368f-faf8-47b9-8c13-4ac72d62ed22)

- **Previous Execution Time (with original index)**:  
  Execution Time: 195.084 ms  
  Planning Time: 0.112 ms  

![Screenshot from 2025-02-17 15-34-31](https://github.com/user-attachments/assets/0e8ac7c3-b3d8-46f7-b999-6fdb4e96a903)

- **Current Execution Time (with optimized index)**:  
  Execution Time: 63.285 ms  
  Planning Time: 0.139 ms 

Note: Please check **Allow edits from maintainers.** if you would like us to assist in the PR.