@Pranav-Kumar-Singh-038 2025-02-17
cc: @tachyons  Could you review this? I have the changes for the performance issue in #5391 .